Added error handling to cache

+++ src/classes/system/cache.class.php
@@ -33,27 +33,31 @@ class Cache {
 			$memc = new Memcached();
 			$memc->addServer('localhost', 11211);
 
-			$memc->delete(SITE_URL."-".$key);
-
+			if($memc->get(SITE_URL."-".$key) ) {
+				$memc->delete(SITE_URL."-".$key);
+			}
 		}
+
 	}
 
-	
 	function unserializeSession($data) {
 
+		$results = "";
 		$vars = preg_split('/([a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*)\|/', $data, -1, PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE);
-		for($i = 0; isset($vars[$i]); $i++) {
-			$results[$vars[$i++]] = unserialize($vars[$i]);
-		}
+		if(count($vars) > 1) {
 
-		foreach($results as $index => $result) {
+			for($i = 0; isset($vars[$i]); $i++) {
+				$results[$vars[$i++]] = unserialize($vars[$i]);
+			}
 
-			$results[$index] = $this->decodeSessionJSON($result);
-			
+			foreach($results as $index => $result) {
+				$results[$index] = $this->decodeSessionJSON($result);
+			}
+
 		}
 		return $results;
 	}
-	
+
 	function decodeSessionJSON($dataset) {
 
 		if(is_array($dataset)) {
